VulnScan-Pro-backend-requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
requests==2.31.0
python-nmap==0.7.1
Werkzeug==2.3.6